Added height changes and some extra folder structure and restructuring.
