+++ railpass/src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import './App.css';
 import { Routes, Route } from "react-router-dom"
-import { Home } from './components/home';
-import { Login } from './components/login';
-import { Register } from './components/register';
-import { Timing } from './components/timing';
+import { Home } from './components/mainPages/home';
+import { Login } from './components/mainPages/login';
+import { Register } from './components/mainPages/register';
+import { Timing } from './components/mainPages/timing';
 
 function App() {
   return (
