+++ railpass/src/components/mainPages/register/index.ts
