+++ railpass/src/components/reusableComponents/titleComponent/index.tsx
@@ -1,0 +1,2 @@
+import Title from "./title";
+export default Title;
