+++ railpass/src/components/reusableComponents/titleComponent/title.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Title(props:any) {
+  return (
+    <h1>{props.title}</h1>
+  )
+}
